+++ src/controllers/BodyParseErrorHandler.ts
@@ -16,16 +16,16 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-import { SETUP_VALIDATION_ERROR_HANDLER, VALIDATION_ERROR_HANDLER } from "../constants";
-import type { InitControllerValidationErrorHandlerAction } from "../types/internal";
+import { PARSE_ERROR_HANDLER, SETUP_PARSE_ERROR_HANDLER } from "../constants";
+import type { InitControllerParseErrorHandlerAction } from "../types/internal";
 import { getListFromObject, getMethodOrThrow } from "./utils";
 
 /**
- * Add a method of a controller as a validation error handler.
+ * Add a method of a controller as a parse error handler.
  *
  * @example
  * ```
- * import { Controller, ControllerBase, IHttpRequest, IHttpResponse, JoiValidationError, POST, schema, ValidationErrorHandler } from '@egomobile/http-server'
+ * import { BodyParseErrorHandler, Controller, ControllerBase, IHttpRequest, IHttpResponse, ParseError, ParseErrorHandler, POST, schema } from '@egomobile/http-server'
  *
  * const mySchema = schema.object({
  *   email: schema.string().trim().email().required(),
@@ -39,9 +39,9 @@ import { getListFromObject, getMethodOrThrow } from "./utils";
  *     response.write('You send: ' + JSON.stringify(response.body!))
  *   }
  *
- *   @ValidationErrorHandler()  // mark that method as default schema validation
- *                              // error handler inside that controller
- *   async handleValidationError(error: JoiValidationError, request: IHttpRequest, response: IHttpResponse) {
+ *   @BodyParseErrorHandler()  // mark that method as default
+ *                             // parse error handler inside that controller
+ *   async handleBodyParseError(error: ParseError, request: IHttpRequest, response: IHttpResponse) {
  *     const errorMessage = Buffer.from('VALIDATION ERROR: ' + error.message, 'utf8')
  *
  *     response.writeHead(400, {
@@ -56,13 +56,13 @@ import { getListFromObject, getMethodOrThrow } from "./utils";
  *
  * @returns {MethodDecorator} The new decorator function.
  */
-export function ValidationErrorHandler(): MethodDecorator {
+export function BodyParseErrorHandler(): MethodDecorator {
     return function (target, methodName, descriptor) {
         const method = getMethodOrThrow(descriptor);
 
-        getListFromObject<InitControllerValidationErrorHandlerAction>(method, SETUP_VALIDATION_ERROR_HANDLER).push(
+        getListFromObject<InitControllerParseErrorHandlerAction>(method, SETUP_PARSE_ERROR_HANDLER).push(
             ({ controller }) => {
-                (controller as any)[VALIDATION_ERROR_HANDLER] = method.bind(controller);
+                (controller as any)[PARSE_ERROR_HANDLER] = method.bind(controller);
             }
         );
     };